# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Management System",
        "version": "1.0.0"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/bookings/create": {
            "post": {
                "tags": [
                    "bookings"
                ],
                "summary": "Create a new booking",
                "operationId": "createBooking",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "roomNum": {
                                        "type": "string"
                                    },
                                    "fromTime": {
                                        "type": "string"
                                    },
                                    "toTime": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "paymentMethod": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "description": "Booking object that needs to be added to the system",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "bookingId": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/bookings/delete": {
            "post": {
                "tags": [
                    "bookings"
                ],
                "summary": "Delete a booking",
                "operationId": "deleteBooking",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "bookingId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "description": "Booking object that needs to be deleted from the system",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "bookingId": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}